add promocodes

+++ app/Models/TourReservation.php
@@ -28,7 +28,7 @@ class TourReservation extends Model
 
     protected $casts = [
         'tour_id' => 'integer',
-        'amount' => 'integer',
+        'amount' => 'double',
         'number_of_pass' => 'integer',
         'order_transaction_id' => 'integer',
         'reserved_user_id' => 'integer'
